% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_block}
\alias{annotate_block}
\title{Annotate block}
\usage{
annotate_block(data, ...)
}
\arguments{
\item{data}{Data to be plotted.}

\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Annotate block for ggplot2.
}
