% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{facet_block}
\alias{facet_block}
\title{Facet block}
\usage{
facet_block(data, ...)
}
\arguments{
\item{data}{Data to be plotted.}

\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Facet wrap block for ggplot2.
}
