% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{geombar_block}
\alias{geombar_block}
\title{Bar block}
\usage{
geombar_block(data, ...)
}
\arguments{
\item{data}{Data to be plotted.}

\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Add bar to a ggplot2 object.
}
