% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{geomboxplot_block}
\alias{geomboxplot_block}
\title{Boxplot block}
\usage{
geomboxplot_block(data, ...)
}
\arguments{
\item{data}{Data to be plotted.}

\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Boxplot block for ggplot2.
}
