% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{geomhistogram_block}
\alias{geomhistogram_block}
\title{Histogram block}
\usage{
geomhistogram_block(data, ...)
}
\arguments{
\item{data}{Data to be plotted.}

\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Add histogram to a ggplot2 object.
}
