% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter.R
\name{geomjitter_block}
\alias{geomjitter_block}
\title{Jitter block}
\usage{
geomjitter_block(data, ...)
}
\arguments{
\item{data}{Data to be plotted.}

\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Jitter block for ggplot2.
}
