% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{geomline_block}
\alias{geomline_block}
\title{Line block}
\usage{
geomline_block(data, ...)
}
\arguments{
\item{data}{Data to be plotted.}

\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Add points to a ggplot2 object.
}
