% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{geompoint_block}
\alias{geompoint_block}
\title{Point block}
\usage{
geompoint_block(data, ...)
}
\arguments{
\item{data}{Data to be plotted.}

\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Add points to a ggplot2 object.
}
