% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot_block}
\alias{ggplot_block}
\title{GGplot block}
\usage{
ggplot_block(data, ...)
}
\arguments{
\item{data}{Data to be plotted.}

\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Ggplot block.
}
