% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{hvline}
\alias{hvline}
\alias{new_geomhline_block}
\alias{new_geomvline_block}
\title{Horinzontal and Vertical Lines block}
\usage{
new_geomhline_block(...)

new_geomvline_block(...)
}
\arguments{
\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Add lines to a ggplot2 object.
}
