% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label_block}
\alias{label_block}
\title{Label block}
\usage{
label_block(data, ...)
}
\arguments{
\item{data}{Data to be plotted.}

\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Add labels to a ggplot2 object.
}
