% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{new_annotate_block}
\alias{new_annotate_block}
\title{Annotate block}
\usage{
new_annotate_block(...)
}
\arguments{
\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Annotate block for ggplot2.
}
