% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{new_facet_block}
\alias{new_facet_block}
\title{Facet block}
\usage{
new_facet_block(...)
}
\arguments{
\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Facet wrap block for ggplot2.
}
