% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{new_geombar_block}
\alias{new_geombar_block}
\title{Bar block}
\usage{
new_geombar_block(...)
}
\arguments{
\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Add bar to a ggplot2 object.
}
