% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{new_geomboxplot_block}
\alias{new_geomboxplot_block}
\title{Boxplot block}
\usage{
new_geomboxplot_block(...)
}
\arguments{
\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Boxplot block for ggplot2.
}
