% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{new_geomhistogram_block}
\alias{new_geomhistogram_block}
\title{Histogram block}
\usage{
new_geomhistogram_block(...)
}
\arguments{
\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Add histogram to a ggplot2 object.
}
