% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter.R
\name{new_geomjitter_block}
\alias{new_geomjitter_block}
\title{Jitter block}
\usage{
new_geomjitter_block(...)
}
\arguments{
\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Jitter block for ggplot2.
}
