% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{new_geomline_block}
\alias{new_geomline_block}
\title{Line block}
\usage{
new_geomline_block(...)
}
\arguments{
\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Add points to a ggplot2 object.
}
