% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{new_geompoint_block}
\alias{new_geompoint_block}
\title{Point block}
\usage{
new_geompoint_block(...)
}
\arguments{
\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Add points to a ggplot2 object.
}
