% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{new_geomsmooth_block}
\alias{new_geomsmooth_block}
\title{Smooth block}
\usage{
new_geomsmooth_block(...)
}
\arguments{
\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Fit a line to a ggplot2 object.
}
