% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{new_ggplot_block}
\alias{new_ggplot_block}
\title{GGplot block}
\usage{
new_ggplot_block(
  func = c("x"),
  default_columns = character(),
  submit = FALSE,
  ...
)
}
\arguments{
\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Ggplot block.
}
