% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{new_label_block}
\alias{new_label_block}
\title{Label block}
\usage{
new_label_block(...)
}
\arguments{
\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Add labels to a ggplot2 object.
}
