% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{new_theme_block}
\alias{new_theme_block}
\title{Theme block}
\usage{
new_theme_block(...)
}
\arguments{
\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Theme a ggplot2 object.
}
