% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_block}
\alias{theme_block}
\title{Theme block}
\usage{
theme_block(data, ...)
}
\arguments{
\item{data}{Data to be plotted.}

\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Theme a ggplot2 object.
}
