% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{hvline}
\alias{hvline}
\alias{geomhline_block}
\alias{geomvline_block}
\title{Horinzontal and Vertical Lines block}
\usage{
geomhline_block(data, ...)

geomvline_block(data, ...)
}
\arguments{
\item{data}{Data to be plotted.}

\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Add lines to a ggplot2 object.
}
